[AIRFLOW-5195] Move GCP Dataflow to core
This commit moves GCP Dataflow from contrib to core.
For more information check AIP-21.
