+++ tests/contrib/utils/test_mlengine_operator_utils.py
@@ -97,8 +97,7 @@ class CreateEvaluateOpsTest(unittest.TestCase):
                 ANY)
             self.assertEqual(success_message['predictionOutput'], result)
 
-        with patch('airflow.contrib.operators.dataflow_operator.'
-                   'DataFlowHook') as mock_dataflow_hook:
+        with patch('airflow.gcp.operators.dataflow.DataFlowHook') as mock_dataflow_hook:
             hook_instance = mock_dataflow_hook.return_value
             hook_instance.start_python_dataflow.return_value = None
             summary.execute(None)
